late-command: Check whether qemu_fw_cfg.ko is present
For releases in flux like the current Ganeti, the kernel ABI change differ
between the kernel d-i was booted with (currently 6.12.33+deb13-amd64)
and the udebs being installed d-i itself (currently 6.12.35+deb13-amd64).

As such, test whether qemu_fw_cfg.ko is actually present, otherwise
the installation fails. This won't work with VMs on routed Ganeti, but
for the interim tests of trixie that's not really relevant.

Bug: T391083

+++ modules/install_server/files/autoinstall/scripts/late_command.sh
@@ -118,9 +118,11 @@ fi
 IFACE=$(ip -4 route list 0/0 | cut -d ' ' -f 5 | head -1)
 
 # Load the qemu module from the guest host
-cp /target/lib/modules/$(uname -r)/kernel/drivers/firmware/qemu_fw_cfg.ko /lib/modules/$(uname -r)/kernel/drivers/firmware/
-depmod
-modprobe qemu_fw_cfg
+if [ -f /target/lib/modules/$(uname -r)/kernel/drivers/firmware/qemu_fw_cfg.ko ]; then
+  cp /target/lib/modules/$(uname -r)/kernel/drivers/firmware/qemu_fw_cfg.ko /lib/modules/$(uname -r)/kernel/drivers/firmware/
+  depmod
+  modprobe qemu_fw_cfg
+fi
 
 # If a qemu VM:
 if [ -f "/sys/firmware/qemu_fw_cfg/by_name/opt/ip6/raw" ]; then
